Automate trusted npm publication
