+++ CHANGELOG.md
@@ -42,8 +42,9 @@ release path fails closed on artifact provenance.
 - Reconcile bursts retain their full bounded backlog instead of discarding events past
   the first batch.
 - Release automation now binds Windows checksums to exact archives, gates GitHub Latest
-  on anonymous install proof and approved, proven npm packages, validates the public
-  Homebrew formula outcome, and attests the immutable daemon image.
+  on anonymous install proof and automatically published, provenance-verified npm
+  packages, validates the public Homebrew formula outcome, and attests the immutable
+  daemon image.
 
 ## [0.2.15] - 2026-07-08
